Yay! It works now.
This program sucks. But it works. It downloads a webpage that has
all of the vocabulary for the JLPT level specified, parses it
into a JSON file, and then spits out ten random words. The words
it has already given are blacklisted, essentially, so that it will
never give you the same two words.